+++ src/main/scala/org/graphsense/Model.scala
@@ -7,3 +7,15 @@ case class ExchangeRatesRaw(
 )
 
 case class ExchangeRates(blockId: Int, fiatValues: Seq[Float])
+
+case class TokenExchangeRatesRaw(
+    asset: String,
+    date: String,
+    fiatValues: Option[Map[String, Float]]
+)
+
+case class TokenExchangeRates(
+    asset: String,
+    blockId: Int,
+    fiatValues: Seq[Float]
+)